refactor : types 디렉토리 정리
types 디렉토리 내 파일은 2가지 종류로 구분되어져 있음
1. DTO 타입
2. 프론트엔드 자체에서 공유하기 위해 작성한 타입

따라서, 1번 타입을 DTO 디렉토리로 2번 타입을 common 디렉토리로 정리하여 용도에 따라 분류하여 사용할 수 있도록 함
