+++ frontend/src/pages/landing/LandingPage.tsx
@@ -1,0 +1,17 @@
+const LandingPage = () => {
+  return (
+    <div className="h-full w-full flex flex-col justify-between">
+      <div className="h-[17.6875rem] w-full shrink-0 flex gap-9">
+        <div className="border w-full"></div>
+        <div className="border w-full"></div>
+      </div>
+      <div className="h-[20.5625rem] w-full shrink-0 flex gap-9">
+        <div className="border w-full"></div>
+        <div className="border w-full"></div>
+        <div className="border w-full"></div>
+      </div>
+    </div>
+  );
+};
+
+export default LandingPage;
